Adds application main class and task for creating submissions.
Spring boot ignores mainClassName if it is set before applying the Spring Boot plug-in. Therefore it cannot be set on the command line using the -P option. Since AutoGrading is just a simple Java executable, a JavaExec task is sufficient for running it.

This allows running the server:
gradlew run
And creating submissions:
gradlew create-submission